Major update, all authentication system is complete

+++ shiftlog-backend/controllers/shiftlogController.js
@@ -5,6 +5,7 @@ const AsyncHandler = require("express-async-handler");
 //@route GET /api/shiftlog
 //@access public
 const getShiftLogs = AsyncHandler(async (req, res) => {
+  // to make it so you can only see your users shifts add this to find function { user_id: req.user.id }
   const shift = await Shift.find();
 
   res.status(200).json(shift);
@@ -23,20 +24,35 @@ const getShiftLog = AsyncHandler(async (req, res) => {
 //@route POST /api/shiftlog
 //@access public
 const createShiftLog = AsyncHandler(async (req, res) => {
-  const { ticket, customer, walkthrough, alarms, notes, date } = req.body;
+  const {
+    ticket,
+    walkthrough,
+    critical_updates,
+    ticket_updates,
+    log_name,
+    date,
+  } = req.body;
 
-  if (!ticket || !customer || !walkthrough || !alarms || !notes || !date) {
+  if (
+    !ticket ||
+    !walkthrough ||
+    !critical_updates ||
+    !ticket_updates ||
+    !log_name ||
+    !date
+  ) {
     res.status(400);
     throw new Error("All fields must be entered");
   }
 
   const log = await Shift.create({
     ticket,
-    customer,
     walkthrough,
-    alarms,
-    notes,
+    critical_updates,
+    ticket_updates,
+    log_name,
     date,
+    user_id: req.user.id,
   });
 
   res.status(200).json(log);
@@ -51,6 +67,11 @@ const updateShiftLog = AsyncHandler(async (req, res) => {
   if (!shift) {
     res.status(400);
     throw new Error("Shift Log Not Found");
+  }
+
+  if (shift.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permission to update other user contacts!");
   }
 
   const updatedShift = await Shift.findByIdAndUpdate(req.params.id, req.body, {
@@ -73,10 +94,10 @@ const deleteShiftLog = AsyncHandler(async (req, res) => {
     throw new Error("Shift not found");
   }
 
-  // // if (shift.user_id.toString() !== req.user.id) {
-  // //   res.status(403);
-  // //   throw new Error("User dont have permission to delete other user contacts!");
-  // // }
+  if (shift.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permission to delete other user contacts!");
+  }
 
   await Shift.deleteOne({ _id: req.params.id });
   res.status(200).json(shift);
